The late push
Converted the core code to the latest main branch code.

The player moves with one joystick and faces the direction it's moving in.

The camera follows smoothly and is a little infront od the player when still

I edited the joystick script slightly so the player rotation wont reset with the joystick

I'll spend the rest of tonight integrating this with networking thing